refactor: extract recordBodyEvent and NormalizeTraceID helpers (M1-3/M1-4)
